+++ src/main/java/com/testvagrant/ekam/commons/io/FileFinder.java
@@ -37,9 +37,7 @@ public class FileFinder {
   }
 
   private String getPath(String path) {
-    return Objects.isNull(path) || path.isEmpty()
-            ? ResourcePaths.ROOT
-            : path;
+    return Objects.isNull(path) || path.isEmpty() ? ResourcePaths.ROOT : path;
   }
 
   public List<File> findWithExtension(String fileExtension) {
@@ -64,7 +62,8 @@ public class FileFinder {
               () ->
                   new RuntimeException(
                       String.format(
-                          "File %s%s not found in %s env and path %s", fileName, fileExtension, env, path)));
+                          "File %s%s not found in %s env and path %s",
+                          fileName, fileExtension, env, path)));
     }
     return fileToSearch;
   }
@@ -80,8 +79,8 @@ public class FileFinder {
       File[] files = rootFile.listFiles();
       assert files != null;
       for (File file : files) {
-        if (!file.isDirectory()) {
-          fileFound = file.getName().endsWith(fileExtensionToSearch);
+        if (!file.isDirectory() || isFileFound(fileExtensionToSearch, file)) {
+          fileFound = isFileFound(fileExtensionToSearch, file);
           if (fileFound) {
             allFiles.add(file);
           }
@@ -92,13 +91,21 @@ public class FileFinder {
     }
   }
 
+  private boolean isFileFound(String fileExtensionToSearch, File file) {
+    return file.getName().endsWith(fileExtensionToSearch);
+  }
+
+  private boolean isFileFound(File file, String fileName, String fileExtensionToSearch) {
+    return file.getName().equals(String.format("%s%s", fileName, fileExtensionToSearch));
+  }
+
   private void collectFile(File rootFile, String fileName, String fileExtensionToSearch) {
     if (rootFile.exists() && rootFile.isDirectory()) {
       File[] files = rootFile.listFiles();
       assert files != null;
       for (File file : files) {
-        if (!file.isDirectory()) {
-          fileFound = file.getName().equals(String.format("%s%s", fileName, fileExtensionToSearch));
+        if (!file.isDirectory() || isFileFound(file, fileName, fileExtensionToSearch)) {
+          fileFound = isFileFound(file, fileName, fileExtensionToSearch);
           if (fileFound) {
             fileToSearch = file;
             break;