+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+# 0.1.0-beta
+* Updated DirectoryFinder to handle ambigious folder names
+* Updated filefinder to handle files like `.app`
+* Updated DataCache to read `cache.size` as a systemproperty
+
 # 0.0.9-beta
 * Reverted io commons version to stable
 * Fixed issue with directory finder when there are conflicting folders